fix: widen telegram_id to BIGINT and harden migration pipeline
Root cause of 422 "Invalid data" on Telegram login: users.telegram_id
was INT (32-bit, max ~2.1B) but Telegram user IDs can exceed 2^31.
MySQL throws DataError which the app handler converts to 422.

Changes:
- User model: telegram_id Integer → BigInteger
- Migration 0019: ALTER TABLE users MODIFY telegram_id BIGINT
- env.py: pre-widen alembic_version.version_num to VARCHAR(128)
  using raw information_schema queries before migrations run
- run-migrations.sh: handle "Data too long for version_num" error
  by widening the column and retrying
- cron/Dockerfile: remove hardcoded repo override, use Alpine defaults,
  clean /etc/periodic to prevent unintended periodic jobs at runtime

+++ backend/app/models/user.py
@@ -1,7 +1,7 @@
 from datetime import datetime, timezone
 from enum import Enum
 
-from sqlalchemy import Boolean, DateTime, Enum as SqlEnum, ForeignKey, Integer, String, false
+from sqlalchemy import BigInteger, Boolean, DateTime, Enum as SqlEnum, ForeignKey, Integer, String, false
 from sqlalchemy.orm import Mapped, mapped_column, relationship
 
 from app.core.database import Base
@@ -17,7 +17,7 @@ class User(Base):
     __tablename__ = "users"
 
     id: Mapped[int] = mapped_column(Integer, primary_key=True, autoincrement=True)
-    telegram_id: Mapped[int] = mapped_column(Integer, unique=True, index=True)
+    telegram_id: Mapped[int] = mapped_column(BigInteger, unique=True, index=True)
     username: Mapped[str | None] = mapped_column(String(255))
     first_name: Mapped[str | None] = mapped_column(String(255))
     last_name: Mapped[str | None] = mapped_column(String(255))
